requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
googletrans==4.0.0-rc1
tensorflow
